add delete func and add numbers to coping table

+++ src/main/java/GetType.java
@@ -65,16 +65,26 @@ public class GetType {
 
     }
 
-    public static void deleteVARTYPETable(Connection conn, Statement stmt) throws SQLException {
 
 
-        // יצירת השאילתה למחיקת הטבלה
-        String deleteTableQuery = "DROP TABLE VARTYPE";
+    public static void deleteVARTYPETable(Connection conn, Statement stmt , ResultSet rs) throws SQLException {
+        try {
+            // SQL query to check if the table exists
+            String checkTableQuery = "SELECT COUNT(*) FROM information_schema.tables WHERE table_name = 'VARTYPE'";
+            rs = stmt.executeQuery(checkTableQuery);
+            rs.next();
+            int tableCount = rs.getInt(1);
 
-        // ביצוע השאילתה
-        stmt.execute(deleteTableQuery);
+            if (tableCount != 0) {
+                // delete the table if it exist
+                String deleteTableQuery = "DROP TABLE VARTYPE";
+                stmt.execute(deleteTableQuery);
+            }
 
-        // סגירת השאילתה והחיבור
+        }
+        catch (SQLException e) {
+            e.printStackTrace();
+        }
     }
 
     public static void createVARTYPETableIfNotExists(Connection conn, Statement stmt, ResultSet rs) {
